XsltSensor - Add XLST sensor which is executed before all other sensors [@Phase(name = Phase.Name.PRE)]. This sensor can be used as pre-scanner step to transform XML files for other sensors. - configuration keys: sonar.cxx.xslt.N.stylesheet; sonar.cxx.xslt.N.inputs; sonar.cxx.xslt.N.outputs; N=1...10 - deprecated: sonar.cxx.other.xslt...
